app/requirements.txt
flask
requests
beautifulsoup4
qbittorrent-api
python-dotenv
transmission-rpc